Add localhost response check in kind-cluster.yml workflow
This update augments the GitHub actions in the `.github/workflows/kind-cluster.yml` file by adding a localhost response check. The new section of the code verifies the correct response from the localhost endpoint, enhancing the robustness and the reliability of the workflow. If the returned HTTP status code is not 404, an error message will be printed and the action will fail.