Initial Commit (#1)
* Initial Commit

* Update the README file of k8s-dev-cluster

Renamed the title and revised the content in README for clearer instructions. The update includes prerequisites, usage guidelines, and contributing protocols. Also added disclaimer to specify the usage of this dev cluster.

* Add logos to k8s-dev-cluster README

Added the logos of `Kind` and `Helm` to the README of the `k8s-dev-cluster`. The images help in enhancing the visual appeal and provide better brand recognition for the tools used in the repository.

* Additionally, updates have been made to the Helmfile.yaml to include comments for each repository and to add a new Cloudflare repository for managing Cloudflare service-related Helm charts.

* Update kind-cluster workflow trigger conditions

The conditions to trigger the 'kind-cluster' workflow have been modified. This workflow now executes only on pull requests on the 'main' branch and push events on the 'master' branch. Furthermore, the setup of Go has been removed from the workflow steps.

* Remove comments from kind-cluster workflow file

The leading comments in the "kind-cluster.yml" file have been removed. This change simplifies the code by eliminating unnecessary comments. Better readability and overall code cleanliness is expected as a result.

* Add lint workflow to k8s dev cluster

A new GitHub Actions workflow was introduced for linting in the k8s-dev-cluster. This configuration will trigger the linting process on "push" and "pull request" events to the master branch. This is hoped to enhance the code quality by enforcing code standards.

* Remove unused KUBERNETES_VERSION and kubeconfig context setup

The KUBERNETES_VERSION variable and the set_kubeconfig_context function have been removed from create_cluster.sh since they were no longer in use. The README.md file was also updated with image placements for better visual presentation of the project. These changes contribute to cleaner, more efficient code.

* The commit message does not fit the code differences provided. According to the code provided, the correct commit message would be:

Change target branch from 'main' to 'master' in kind-cluster.yml

In the GitHub workflows for the k8s-dev-cluster project, the target branch for pull requests has been changed from 'main' to 'master'. This change aligns with the project's branching strategy and ensures that pull requests are correctly targeted.

* Add helmfile installation and application to k8s workflow

The kind-cluster workflow within GitHub workflows has been updated to include the steps of installing and applying a helmfile. The README file was also updated to reflect the proper Helmfile repository link. This addition ensures helmfile's availability and improves the workflow's functionality in the k8s-dev-cluster project.

* Also, in the kind-cluster.yml workflow file, the helmfile download URL was corrected from darwin to linux version, and added file permission change step to make helmfile executable after being moved to /usr/local/bin directory.

* Improve helmfile sync retry mechanism in kind-cluster.yml

The updated code in the `.github/workflows/kind-cluster.yml` file now includes a retry mechanism for the `helmfile sync` command. This helps to mitigate issues where the initial execution of the command fails, by attempting to run it twice more before exiting the script with an error.

* Add localhost response check in kind-cluster.yml workflow

This update augments the GitHub actions in the `.github/workflows/kind-cluster.yml` file by adding a localhost response check. The new section of the code verifies the correct response from the localhost endpoint, enhancing the robustness and the reliability of the workflow. If the returned HTTP status code is not 404, an error message will be printed and the action will fail.

* Correct arithmetic syntax in kind-cluster.yml workflow

This commit adjusts the syntax for an arithmetic operation in the `.github/workflows/kind-cluster.yml` file. Previously, the code utilized the deprecated syntax `$[$command_retry+1]`, but this has been updated to `$((command_retry+1))` which provides more predictable results and compatibility in increasing the `command_retry` count.

* Update kubeconfig reference in create_cluster.sh script

This commit adjusts the kubeconfig reference in the `create_cluster.sh` script from `$KUBECONFIG` to `"$KUBECONFIG"`. This modification in the syntax ensures that the value of the variable KUBECONFIG is properly quoted, preventing potential issues arising from spaces or special characters in the value.